增加源文件模块

+++ src/module/error/error.cpp
@@ -3,7 +3,7 @@
 
 const char error_html[] = "<!DOCTYPE html><html><head><meta http-equiv=\"Content-type\" content=\"text/html;charset=UTF-8\"><title>大作业-错误</title></head><body><h1>出错啦！</h1><p>我们找不到请求的文件或命令，请返回到上一页。</p></body></html>";
 
-int error_response::length()
+long long error_response::length()
 {
     return sizeof(error_html) - 1;
 }
