+++ src/module/error/error.h
@@ -7,7 +7,7 @@ public:
     virtual ~error_response() {}
     virtual bool supports(std::string) { return true; }
 
-    virtual int length();
+    virtual long long length();
     virtual int status() { return 404; }
     virtual ssize_t send(int fd);
 };
